+++ CHANGELOG.md
@@ -58,7 +58,8 @@
 
 ### 功能移除
 
-- (无)
+- 去除支持实体维护服务的重置功能。
+  - com.dwarfeng.scg.stack.service.GeneratorSupportMaintainService。
 
 ---